// src/TS/global.d.ts
export interface iconsType {
  id: number;
  src: string;
  alt: string;
  text: string;
}
export interface storySliderType {
  id: number;
  src: string;
  alt: string;
  text: string;
}
export interface categorizeShoppingType {
  id: number;
  src: string;
  alt: string;
  text: string;
}
export interface clientTasteType {
  id: number;
  category: string;
  src1: string;
  src2: string;
  src3: string;
  src4: string;
}
export interface mostPopularBrandsType {
  id: number;
  src: string;
  alt: string;
}
export interface selectedOfferType {
  id: number;
  src: string;
  price: string;
  offPercent: string;
  purePrice: string;
}
export interface partnerShipType {
  id: number;
  src: string;
  alt: string;
}
export interface readingBlogType {
  id: number;
  src: string;
  alt: string;
  text: string;
}

export interface MostSellType {
  id: number;
  src: string;
  alt: string;
  count: string;
  text: string;
  src2: string;
  alt2: string;
  count2: string;
  text2: string;
  src3: string;
  alt3: string;
  count3: string;
  text3: string;
}

export interface MainSliderType {
  id: number;
  src: string;
  alt: string;
}

export interface SuperMarketType {
  id: number;
  src: string;
  alt: string;
  offPercent: string;
}

export interface FeaturesType {
  id: number;
  src: string;
  alt: string;
  text: string;
}

export interface WithDigikalaType {
  id: number;
  text: string;
  href: string;
}
export interface CustomerType {
  id: number;
  text: string;
  href: string;
}

export interface GuideType {
  id: number;
  text: string;
  href: string;
}

export interface DigiChildType {
  id: number;
  src: string;
  alt: string;
}

export interface AmazingSliderType {
  id: number;
  src: string;
  alt: string;
  text: string;
  offPercent: string;
  price: string;
  purePrice: string;
}

export interface PopularCityType {
  id: number;
  text: string;
}

export interface AllCityType {
  id: number;
  area: string;
  city: string;
}
export interface HighlightDigiServiceType {
  id: number;
  src: string;
  text: string;
}

export interface DigiServeseTypec {
  id: number;
  src: string;
  text: string;
}
export interface MegaMenuItemType {
  id: number;
  text: string;
  src: string;
  srcHover: string;
  subMenu: SubMenuType[];
}
export interface SubMenuType {
  id: number;
  boldtext: string;
  items: SubMenuItemType[];
}
export interface SubMenuItemType {
  id: number;
  Subtext: string;
}
export interface SearchBarItemsType {
  id: number;
  text: string;
}